+++ Hermit.Unity/Vender/UIStack/Core/Abstractions/IUIStack.cs
@@ -5,13 +5,13 @@ namespace Hermit.UIStack
 {
     public interface IUIStack : IUIController
     {
-        Task<int> PushAsync(string widgetName);
+        Task<ulong> PushAsync(string widgetName);
 
-        Task<int> PushAsync(string widgetName, UIMessage message);
+        Task<ulong> PushAsync(string widgetName, UIMessage message);
 
-        Task<int> PushAsync<TWidget>(string widgetName) where TWidget : Widget;
+        Task<ulong> PushAsync<TWidget>(string widgetName) where TWidget : Widget;
 
-        Task<int> PushAsync<TWidget>(string widgetName, UIMessage message) where TWidget : Widget;
+        Task<ulong> PushAsync<TWidget>(string widgetName, UIMessage message) where TWidget : Widget;
 
         Task PopAsync(bool recycle = false);
